(MODULES-5493) Manage Two Same Name Appications
This commit fixes a bug in the this module uses the composite name key
feature. Prior to this change the `sitename` property was specified as
a part of the composite key. Unfortunately, properties are not valid
components of a compsite key and the directive was ignored. This meant
that in truth there was only attribute considered for the naming key,
and thus for uniqueness, which was the `applicationname` parameter.

This commit changes the `sitename` property to a paremeter to ensure
that it is considered a part of the composite key. Now if you have two
sites that need an application with the same name on a single instance
of IIS, you can do that.